Added CustomerStaffController.cs
